requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aistudio-sdk==0.3.7
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
Authlib==1.6.1
backcall==0.2.0
backoff==2.2.1
bce-python-sdk==0.9.45
bcrypt==4.3.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
build==1.2.2.post1
cachetools==5.5.2
camelot-py==1.0.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.0.15
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
durationpy==0.10
effdet==0.4.1
einops==0.8.1
emoji==2.14.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.55.7
frozenlist==1.7.0
fsspec==2024.6.1
future==1.0.0
fuzzywuzzy==0.18.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.7.0
google-api-core==1.34.1
google-auth==2.40.3
google-cloud-vision==1.0.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.74.0
grpcio-health-checking==1.74.0
grpcio-status==1.49.0rc1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.8
language_data==1.3.0
layoutparser==0.3.4
Levenshtein==0.26.1
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mkl==2021.4.0
mmh3==5.1.0
modelscope==1.29.2
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.6.3
murmurhash==1.0.12
mypy_extensions==1.1.0
narwhals==1.46.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notion-client==2.4.0
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
ollama==0.5.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt-einsum==3.3.0
orjson==3.11.0
overrides==7.7.0
packaging==24.2
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2023.2.7
pi_heif==1.0.0
pickleshare==0.7.5
pikepdf==9.10.2
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.6
poppler-utils==0.1.0
portalocker==3.2.0
posthog==5.4.0
preshed==3.0.9
prettytable==3.16.0
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==6.1.1
PuLP==3.2.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pymongo==4.13.2
PyMuPDF==1.26.3
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.9.0
pypdfium2==4.30.1
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
PyQRCode==1.2.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.1
rapidocr-onnxruntime==1.4.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
robotframework==7.3.2
robotframework-robosapiens==2.20.9
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
segtok==1.5.11
sentence-transformers==5.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
streamlit==1.46.1
sympy==1.13.3
tabulate==0.9.0
tbb==2021.11.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.11.0
unstructured==0.15.14
unstructured-client==0.41.0
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.15
urllib3==2.3.0
uvicorn==0.35.0
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==4.16.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
yacs==0.1.8
yake==0.4.8
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
